fixed mobile menu header bug (h-full -> h-screen) and refactored mobile header to Layout
